Move tide functions to be closer to where structs were defined
