+++ prow/tide/codereview.go
@@ -19,13 +19,9 @@ package tide
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
-	"strconv"
-	"sync"
 	"time"
 
 	"github.com/sirupsen/logrus"
-	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	prowapi "k8s.io/test-infra/prow/apis/prowjobs/v1"
 	"k8s.io/test-infra/prow/config"
 	"k8s.io/test-infra/prow/git/types"
@@ -192,134 +188,4 @@ type provider interface {
 	GetTideContextPolicy(gitClient git.ClientFactory, org, repo, branch string, baseSHAGetter config.RefGetter, headSHA string) (contextChecker, error)
 	refsForJob(sp subpool, prs []CodeReviewCommon) prowapi.Refs
 	prMergeMethod(crc *CodeReviewCommon) (types.PullRequestMergeType, error)
-}
-
-// GitHubProvider implements provider, used by tide Controller for
-// interacting directly with GitHub.
-//
-// Tide Controller should only use GitHubProvider for communicating with GitHub.
-type GitHubProvider struct {
-	cfg                config.Getter
-	ghc                githubClient
-	usesGitHubAppsAuth bool
-
-	*mergeChecker
-	logger *logrus.Entry
-}
-
-func newGitHubProvider(
-	logger *logrus.Entry,
-	ghc githubClient,
-	cfg config.Getter,
-	mergeChecker *mergeChecker,
-	usesGitHubAppsAuth bool,
-) *GitHubProvider {
-	return &GitHubProvider{
-		logger:             logger,
-		ghc:                ghc,
-		cfg:                cfg,
-		usesGitHubAppsAuth: usesGitHubAppsAuth,
-		mergeChecker:       mergeChecker,
-	}
-}
-
-func (gi *GitHubProvider) blockers() (blockers.Blockers, error) {
-	label := gi.cfg().Tide.BlockerLabel
-	if label == "" {
-		return blockers.Blockers{}, nil
-	}
-
-	gi.logger.WithField("blocker_label", label).Debug("Searching for blocker issues")
-	orgExcepts, repos := gi.cfg().Tide.Queries.OrgExceptionsAndRepos()
-	orgs := make([]string, 0, len(orgExcepts))
-	for org := range orgExcepts {
-		orgs = append(orgs, org)
-	}
-	orgRepoQuery := orgRepoQueryStrings(orgs, repos.UnsortedList(), orgExcepts)
-	return blockers.FindAll(gi.ghc, gi.logger, label, orgRepoQuery, gi.usesGitHubAppsAuth)
-}
-
-// Query gets all open PRs based on tide configuration.
-func (gi *GitHubProvider) Query() (map[string]CodeReviewCommon, error) {
-	lock := sync.Mutex{}
-	wg := sync.WaitGroup{}
-	prs := make(map[string]CodeReviewCommon)
-	var errs []error
-	for i, query := range gi.cfg().Tide.Queries {
-
-		// Use org-sharded queries only when GitHub apps auth is in use
-		var queries map[string]string
-		if gi.usesGitHubAppsAuth {
-			queries = query.OrgQueries()
-		} else {
-			queries = map[string]string{"": query.Query()}
-		}
-
-		for org, q := range queries {
-			org, q, i := org, q, i
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				results, err := gi.search(gi.ghc.QueryWithGitHubAppsSupport, gi.logger, q, time.Time{}, time.Now(), org)
-
-				resultString := "success"
-				if err != nil {
-					resultString = "error"
-				}
-				tideMetrics.queryResults.WithLabelValues(strconv.Itoa(i), org, resultString).Inc()
-
-				lock.Lock()
-				defer lock.Unlock()
-				if err != nil && len(results) == 0 {
-					gi.logger.WithField("query", q).WithError(err).Warn("Failed to execute query.")
-					errs = append(errs, fmt.Errorf("query %d, err: %w", i, err))
-					return
-				}
-				if err != nil {
-					gi.logger.WithError(err).WithField("query", q).Warning("found partial results")
-				}
-
-				for _, pr := range results {
-					crc := CodeReviewCommonFromPullRequest(&pr)
-					prs[prKey(crc)] = *crc
-				}
-			}()
-		}
-	}
-	wg.Wait()
-
-	return prs, utilerrors.NewAggregate(errs)
-}
-
-func (gi *GitHubProvider) GetRef(org, repo, ref string) (string, error) {
-	return gi.ghc.GetRef(org, repo, ref)
-}
-
-func (gi *GitHubProvider) GetTideContextPolicy(gitClient git.ClientFactory, org, repo, branch string, baseSHAGetter config.RefGetter, headSHA string) (contextChecker, error) {
-	return gi.cfg().GetTideContextPolicy(gitClient, org, repo, branch, baseSHAGetter, headSHA)
-}
-
-func (gi *GitHubProvider) refsForJob(sp subpool, prs []CodeReviewCommon) prowapi.Refs {
-	refs := prowapi.Refs{
-		Org:     sp.org,
-		Repo:    sp.repo,
-		BaseRef: sp.branch,
-		BaseSHA: sp.sha,
-	}
-	for _, pr := range prs {
-		refs.Pulls = append(
-			refs.Pulls,
-			prowapi.Pull{
-				Number: pr.Number,
-				Title:  pr.Title,
-				Author: string(pr.AuthorLogin),
-				SHA:    pr.HeadRefOID,
-			},
-		)
-	}
-	return refs
-}
-
-func (gi *GitHubProvider) prMergeMethod(crc *CodeReviewCommon) (types.PullRequestMergeType, error) {
-	return gi.mergeChecker.prMergeMethod(gi.cfg().Tide, crc)
 }
