--- prow/tide/search.go
@@ -1,106 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package tide
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"k8s.io/test-infra/prow/github"
-
-	githubql "github.com/shurcooL/githubv4"
-	"github.com/sirupsen/logrus"
-)
-
-type querier func(ctx context.Context, q interface{}, vars map[string]interface{}, org string) error
-
-func datedQuery(q string, start, end time.Time) string {
-	return fmt.Sprintf("%s %s", q, dateToken(start, end))
-}
-
-func floor(t time.Time) time.Time {
-	if t.Before(github.FoundingYear) {
-		return github.FoundingYear
-	}
-	return t
-}
-
-func (gi *GitHubProvider) search(query querier, log *logrus.Entry, q string, start, end time.Time, org string) ([]PullRequest, error) {
-	start = floor(start)
-	end = floor(end)
-	log = log.WithFields(logrus.Fields{
-		"query": q,
-		"start": start.String(),
-		"end":   end.String(),
-	})
-	requestStart := time.Now()
-	var cursor *githubql.String
-	vars := map[string]interface{}{
-		"query":        githubql.String(datedQuery(q, start, end)),
-		"searchCursor": cursor,
-	}
-
-	var totalCost, remaining int
-	var ret []PullRequest
-	var sq searchQuery
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	defer cancel()
-	for {
-		log.Debug("Sending query")
-		if err := query(ctx, &sq, vars, org); err != nil {
-			if cursor != nil {
-				err = fmt.Errorf("cursor: %q, err: %w", *cursor, err)
-			}
-			return ret, err
-		}
-		totalCost += int(sq.RateLimit.Cost)
-		remaining = int(sq.RateLimit.Remaining)
-		for _, n := range sq.Search.Nodes {
-			ret = append(ret, n.PullRequest)
-		}
-		if !sq.Search.PageInfo.HasNextPage {
-			break
-		}
-		cursor = &sq.Search.PageInfo.EndCursor
-		vars["searchCursor"] = cursor
-		log = log.WithField("searchCursor", *cursor)
-	}
-	log.WithFields(logrus.Fields{
-		"duration":       time.Since(requestStart).String(),
-		"pr_found_count": len(ret),
-		"cost":           totalCost,
-		"remaining":      remaining,
-	}).Debug("Finished query")
-	return ret, nil
-}
-
-// dateToken generates a GitHub search query token for the specified date range.
-// See: https://help.github.com/articles/understanding-the-search-syntax/#query-for-dates
-func dateToken(start, end time.Time) string {
-	// GitHub's GraphQL API silently fails if you provide it with an invalid time
-	// string.
-	// Dates before 1970 (unix epoch) are considered invalid.
-	startString, endString := "*", "*"
-	if start.Year() >= 1970 {
-		startString = start.Format(github.SearchTimeFormat)
-	}
-	if end.Year() >= 1970 {
-		endString = end.Format(github.SearchTimeFormat)
-	}
-	return fmt.Sprintf("updated:%s..%s", startString, endString)
-}
