+++ prow/tide/tide_test.go
@@ -2057,103 +2057,6 @@ func TestServeHTTP(t *testing.T) {
 	}
 }
 
-func TestHeadContexts(t *testing.T) {
-	type commitContext struct {
-		// one context per commit for testing
-		context string
-		sha     string
-	}
-
-	win := "win"
-	lose := "lose"
-	headSHA := "head"
-	testCases := []struct {
-		name                string
-		commitContexts      []commitContext
-		expectAPICall       bool
-		expectChecksAPICall bool
-	}{
-		{
-			name: "first commit is head",
-			commitContexts: []commitContext{
-				{context: win, sha: headSHA},
-				{context: lose, sha: "other"},
-				{context: lose, sha: "sha"},
-			},
-		},
-		{
-			name: "last commit is head",
-			commitContexts: []commitContext{
-				{context: lose, sha: "shaaa"},
-				{context: lose, sha: "other"},
-				{context: win, sha: headSHA},
-			},
-		},
-		{
-			name: "no commit is head, falling back to v3 api and getting context via status api",
-			commitContexts: []commitContext{
-				{context: lose, sha: "shaaa"},
-				{context: lose, sha: "other"},
-				{context: lose, sha: "sha"},
-			},
-			expectAPICall: true,
-		},
-		{
-			name: "no commit is head, falling back to v3 api and getting context via checks api",
-			commitContexts: []commitContext{
-				{context: lose, sha: "shaaa"},
-				{context: lose, sha: "other"},
-				{context: lose, sha: "sha"},
-			},
-			expectAPICall:       true,
-			expectChecksAPICall: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Logf("Running test case %q", tc.name)
-			fgc := &fgc{}
-			if !tc.expectChecksAPICall {
-				fgc.combinedStatus = map[string]string{win: string(githubql.StatusStateSuccess)}
-			} else {
-				fgc.checkRuns = &github.CheckRunList{CheckRuns: []github.CheckRun{
-					{Name: win, Status: "completed", Conclusion: "neutral"},
-				}}
-			}
-			if tc.expectAPICall {
-				fgc.expectedSHA = headSHA
-			}
-			provider := &GitHubProvider{
-				ghc:    fgc,
-				logger: logrus.WithField("component", "tide"),
-			}
-			pr := &PullRequest{HeadRefOID: githubql.String(headSHA)}
-			for _, ctx := range tc.commitContexts {
-				commit := Commit{
-					Status: struct{ Contexts []Context }{
-						Contexts: []Context{
-							{
-								Context: githubql.String(ctx.context),
-							},
-						},
-					},
-					OID: githubql.String(ctx.sha),
-				}
-				pr.Commits.Nodes = append(pr.Commits.Nodes, struct{ Commit Commit }{commit})
-			}
-
-			contexts, err := provider.headContexts(CodeReviewCommonFromPullRequest(pr))
-			if err != nil {
-				t.Fatalf("Unexpected error from headContexts: %v", err)
-			}
-			if len(contexts) != 1 || string(contexts[0].Context) != win {
-				t.Errorf("Expected exactly 1 %q context, but got: %#v", win, contexts)
-			}
-		})
-	}
-}
-
 func testPR(org, repo, branch string, number int, mergeable githubql.MergeableState) *PullRequest {
 	pr := PullRequest{
 		Number:     githubql.Int(number),
@@ -3263,104 +3166,6 @@ func TestPresubmitsByPull(t *testing.T) {
 func getTemplate(name, tplStr string) *template.Template {
 	tpl, _ := template.New(name).Parse(tplStr)
 	return tpl
-}
-
-func TestPrepareMergeDetails(t *testing.T) {
-	pr := PullRequest{
-		Number:     githubql.Int(1),
-		Mergeable:  githubql.MergeableStateMergeable,
-		HeadRefOID: githubql.String("SHA"),
-		Title:      "my commit title",
-		Body:       "my commit body",
-	}
-
-	testCases := []struct {
-		name        string
-		tpl         config.TideMergeCommitTemplate
-		pr          PullRequest
-		mergeMethod types.PullRequestMergeType
-		expected    github.MergeDetails
-	}{{
-		name:        "No commit template",
-		tpl:         config.TideMergeCommitTemplate{},
-		pr:          pr,
-		mergeMethod: "merge",
-		expected: github.MergeDetails{
-			SHA:         "SHA",
-			MergeMethod: "merge",
-		},
-	}, {
-		name: "No commit template fields",
-		tpl: config.TideMergeCommitTemplate{
-			Title: nil,
-			Body:  nil,
-		},
-		pr:          pr,
-		mergeMethod: "merge",
-		expected: github.MergeDetails{
-			SHA:         "SHA",
-			MergeMethod: "merge",
-		},
-	}, {
-		name: "Static commit template",
-		tpl: config.TideMergeCommitTemplate{
-			Title: getTemplate("CommitTitle", "static title"),
-			Body:  getTemplate("CommitBody", "static body"),
-		},
-		pr:          pr,
-		mergeMethod: "merge",
-		expected: github.MergeDetails{
-			SHA:           "SHA",
-			MergeMethod:   "merge",
-			CommitTitle:   "static title",
-			CommitMessage: "static body",
-		},
-	}, {
-		name: "Commit template uses PullRequest fields",
-		tpl: config.TideMergeCommitTemplate{
-			Title: getTemplate("CommitTitle", "{{ .Number }}: {{ .Title }}"),
-			Body:  getTemplate("CommitBody", "{{ .HeadRefOID }} - {{ .Body }}"),
-		},
-		pr:          pr,
-		mergeMethod: "merge",
-		expected: github.MergeDetails{
-			SHA:           "SHA",
-			MergeMethod:   "merge",
-			CommitTitle:   "1: my commit title",
-			CommitMessage: "SHA - my commit body",
-		},
-	}, {
-		name: "Commit template uses nonexistent fields",
-		tpl: config.TideMergeCommitTemplate{
-			Title: getTemplate("CommitTitle", "{{ .Hello }}"),
-			Body:  getTemplate("CommitBody", "{{ .World }}"),
-		},
-		pr:          pr,
-		mergeMethod: "merge",
-		expected: github.MergeDetails{
-			SHA:         "SHA",
-			MergeMethod: "merge",
-		},
-	}}
-
-	for _, test := range testCases {
-		t.Run(test.name, func(t *testing.T) {
-			cfg := &config.Config{}
-			cfgAgent := &config.Agent{}
-			cfgAgent.Set(cfg)
-			provider := &GitHubProvider{
-				cfg:    cfgAgent.Config,
-				ghc:    &fgc{},
-				logger: logrus.WithContext(context.Background()),
-			}
-
-			actual := provider.prepareMergeDetails(test.tpl, *CodeReviewCommonFromPullRequest(&test.pr), test.mergeMethod)
-
-			if !reflect.DeepEqual(actual, test.expected) {
-				t.Errorf("Case %s failed: expected %+v, got %+v", test.name, test.expected, actual)
-			}
-		})
-	}
 }
 
 func TestAccumulateReturnsCorrectMissingTests(t *testing.T) {
